requirements.txt
ipykernel
tqdm
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
langchain-huggingface
faiss-cpu
langchain-google-genai
pypdf
hf_xet
langchain-pinecone
langchain_experimental
fastembed
sentence-transformers
matplotlib
ipywidgets
unstructured[all-docs]
pytesseract
pillow
lxml
imagehash
clip
pyvips
opensearch-py
langchain-community
langchain_milvus